fix(angular): fix incremental builds with angular 13
