+++ packages/angular/src/executors/ng-packagr-lite/ng-packagr-adjustments/ng-package/entry-point/compile-ngc.transform.ts
@@ -5,27 +5,18 @@
  * since these libraries will be compiled by the ngtsc.
  */
 
-import { InjectionToken, Provider } from 'injection-js';
 import type { Transform } from 'ng-packagr/lib/graph/transform';
 import { transformFromPromise } from 'ng-packagr/lib/graph/transform';
-import { provideTransform } from 'ng-packagr/lib/graph/transform.di';
 import {
   isEntryPoint,
   isEntryPointInProgress,
 } from 'ng-packagr/lib/ng-package/nodes';
-import {
-  NgPackagrOptions,
-  OPTIONS_TOKEN,
-} from 'ng-packagr/lib/ng-package/options.di';
-import { compileSourceFiles } from './compile-source-files';
+import { NgPackagrOptions } from 'ng-packagr/lib/ng-package/options.di';
 import type { StylesheetProcessor as StylesheetProcessorClass } from 'ng-packagr/lib/styles/stylesheet-processor';
-import {
-  STYLESHEET_PROCESSOR,
-  STYLESHEET_PROCESSOR_TOKEN,
-} from 'ng-packagr/lib/styles/stylesheet-processor.di';
 import { setDependenciesTsConfigPaths } from 'ng-packagr/lib/ts/tsconfig';
 import * as path from 'path';
 import * as ts from 'typescript';
+import { compileSourceFiles } from '../../ngc/compile-source-files';
 
 export const nxCompileNgcTransformFactory = (
   StylesheetProcessor: typeof StylesheetProcessorClass,
@@ -75,16 +66,3 @@ export const nxCompileNgcTransformFactory = (
     return graph;
   });
 };
-
-export const NX_COMPILE_NGC_TOKEN = new InjectionToken<Transform>(
-  `nx.v1.compileNgc`
-);
-export const NX_COMPILE_NGC_TRANSFORM = provideTransform({
-  provide: NX_COMPILE_NGC_TOKEN,
-  useFactory: nxCompileNgcTransformFactory,
-  deps: [STYLESHEET_PROCESSOR_TOKEN, OPTIONS_TOKEN],
-});
-export const NX_COMPILE_NGC_PROVIDERS: Provider[] = [
-  STYLESHEET_PROCESSOR,
-  NX_COMPILE_NGC_TRANSFORM,
-];
