+++ packages/angular/src/executors/ng-packagr-lite/ng-packagr-adjustments/ng-package/entry-point/entry-point.di.ts
@@ -1,15 +1,22 @@
 /**
- * Wires everything together and provides it to the DI system, taking what
- * we still want from the original ng-packagr library and replacing those
- * where Nx takes over with Nx specific functions
+ * Adapted from the original ng-packagr source.
+ *
+ * Changes made:
+ * - Provide our own entryPointTransformFactory function.
+ * - Use NX_COMPILE_NGC_TOKEN instead of COMPILE_NGC_TOKEN.
+ * - Use NX_COMPILE_NGC_PROVIDERS instead of COMPILE_NGC_PROVIDERS.
+ * - Removed usage of WRITE_BUNDLES_TRANSFORM_TOKEN and WRITE_BUNDLES_TRANSFORM.
  */
 
 import type { Provider } from 'injection-js';
 import { InjectionToken } from 'injection-js';
 import type { Transform } from 'ng-packagr/lib/graph/transform';
 import { provideTransform } from 'ng-packagr/lib/graph/transform.di';
-import { NX_COMPILE_NGC_PROVIDERS, NX_COMPILE_NGC_TOKEN } from './compile-ngc';
-import { nxEntryPointTransformFactory } from './entry-point';
+import {
+  NX_COMPILE_NGC_PROVIDERS,
+  NX_COMPILE_NGC_TOKEN,
+} from './compile-ngc.di';
+import { nxEntryPointTransformFactory } from './entry-point.transform';
 import {
   NX_WRITE_PACKAGE_TRANSFORM,
   NX_WRITE_PACKAGE_TRANSFORM_TOKEN,
